+++ packages/lumx-react/src/components/thumbnail/Thumbnail.stories.tsx
@@ -1,8 +1,13 @@
 import { Alignment, AspectRatio, CrossOrigin, Size, Theme, Thumbnail, ThumbnailVariant } from '@lumx/react';
+import { htmlDecode } from '@lumx/react/utils/htmlDecode';
 import { boolean, number, select, text } from '@storybook/addon-knobs';
 import React from 'react';
 
-export default { title: 'LumX components/Thumbnail' };
+export default { title: 'LumX components/thumbnail/Thumbnail' };
+
+interface StoryProps {
+    theme: Theme;
+}
 
 const numberKnobOptions = {
     max: 1,
@@ -11,30 +16,52 @@ const numberKnobOptions = {
     step: 0.1,
 };
 
+const corsOptions = {
+    None: undefined,
+    Anonymous: CrossOrigin.anonymous,
+    UseCredentials: CrossOrigin.useCredentials,
+};
+
+const groupId = 'Options';
+
 /**
  * Thumbnail story
  * @return simple Thumbnail.
  */
-export const defaultThumbnail = ({ theme }: { theme: Theme }) => {
+export const defaultThumbnail = ({ theme }: StoryProps) => {
+    const align = select<Alignment>('Alignment', Alignment, Alignment.left, groupId);
+    const aspectRatio = select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, groupId);
+    const isCrossOriginEnabled = boolean('Is CORS enabled', false, groupId);
+    const crossOrigin = select('CORS', corsOptions, corsOptions.None, groupId);
+    const fillHeight = boolean('Fill Height', false, groupId);
+    const focusPoint = {
+        x: number('focusX', 0, numberKnobOptions, groupId),
+        y: number('focusY', 0, numberKnobOptions, groupId),
+    };
+    const image = text(
+        'Url',
+        'https://i.picsum.photos/id/237/536/354.jpg?hmac=i0yVXW1ORpyCZpQ-CknuyV-jbtU7_x9EBQVhvT5aRr0',
+        groupId,
+    );
+    const isFollowingWindowSize = boolean('Update on window resize', true, groupId);
+    const resizeDebounceTime = number('Debounce time after resize', 20, undefined, groupId);
+    const variant = select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, groupId);
+
     return (
         <Thumbnail
-            align={select<Alignment>('Alignment', Alignment, Alignment.left, 'Options')}
-            aspectRatio={select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, 'Options')}
-            isCrossOriginEnabled={boolean('Enable CORS', true, 'Options')}
-            crossOrigin={select('CORS', CrossOrigin, CrossOrigin.anonymous, 'Options')}
-            fillHeight={boolean('Fill Height', false, 'Options')}
-            focusPoint={{
-                x: number('focusX', 0, numberKnobOptions, 'Options'),
-                y: number('focusY', 0, numberKnobOptions, 'Options'),
-            }}
-            image={text(
-                'Url image',
-                'https://i.picsum.photos/id/237/536/354.jpg?hmac=i0yVXW1ORpyCZpQ-CknuyV-jbtU7_x9EBQVhvT5aRr0',
-                'Options',
-            )}
+            align={align}
+            aspectRatio={aspectRatio}
+            isCrossOriginEnabled={isCrossOriginEnabled}
+            crossOrigin={crossOrigin}
+            fillHeight={fillHeight}
+            focusPoint={focusPoint}
+            image={htmlDecode(image)}
+            isFollowingWindowSize={isFollowingWindowSize}
+            resizeDebounceTime={resizeDebounceTime}
             size={select(
                 'Size',
                 {
+                    None: undefined,
                     XXS: Size.xxs,
                     XS: Size.xs,
                     S: Size.s,
@@ -44,29 +71,91 @@ export const defaultThumbnail = ({ theme }: { theme: Theme }) => {
                     XXL: Size.xxl,
                 },
                 Size.xxl,
-                'Options',
+                groupId,
             )}
             theme={theme}
-            isFollowingWindowSize={boolean('Update on window resize', true, 'Options')}
-            resizeDebounceTime={number('Debounce time after resize', 20, undefined, 'Options')}
-            variant={select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, 'Options')}
+            variant={variant}
         />
     );
 };
 
-export const brokenThumbnailWithFallbackProps = ({ theme }: { theme: Theme }) => (
-    <Thumbnail
-        align={select<Alignment>('Alignment', Alignment, Alignment.left, 'Options')}
-        aspectRatio={select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, 'Options')}
-        isCrossOriginEnabled={boolean('Enable CORS', true, 'Options')}
-        crossOrigin={select('CORS', CrossOrigin, CrossOrigin.anonymous, 'Options')}
-        fillHeight={boolean('Fill Height', false, 'Options')}
-        image={text(
-            'Url image',
-            'https://i.picsum.photos/id/237/536/354.jp?hmac=i0yVXW1ORpyCZpQ-CknuyV-jbtU7_x9EBQVhvT5aRr0',
-            'Options',
-        )}
-        theme={theme}
-        variant={select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, 'Options')}
-    />
-);
+export const thumbnailWithCustomImgProps = ({ theme }: StoryProps) => {
+    const [image, setImage] = React.useState('https://not-found');
+    const onError = () => {
+        // tslint:disable-next-line: no-console
+        console.log('onError event called');
+        setImage('https://i.picsum.photos/id/237/536/354.jpg?hmac=i0yVXW1ORpyCZpQ-CknuyV-jbtU7_x9EBQVhvT5aRr0');
+    };
+
+    const align = select<Alignment>('Alignment', Alignment, Alignment.left, groupId);
+    const aspectRatio = select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, groupId);
+    const isCrossOriginEnabled = boolean('Is CORS enabled', false, groupId);
+    const crossOrigin = select('CORS', corsOptions, corsOptions.Anonymous, groupId);
+    const fillHeight = boolean('Fill Height', false, groupId);
+    const focusPoint = {
+        x: number('focusX', 0, numberKnobOptions, groupId),
+        y: number('focusY', 0, numberKnobOptions, groupId),
+    };
+    const isFollowingWindowSize = boolean('Update on window resize', true, groupId);
+    const resizeDebounceTime = number('Debounce time after resize', 20, undefined, groupId);
+    const variant = select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, groupId);
+
+    return (
+        <Thumbnail
+            align={align}
+            aspectRatio={aspectRatio}
+            isCrossOriginEnabled={isCrossOriginEnabled}
+            crossOrigin={crossOrigin}
+            fillHeight={fillHeight}
+            focusPoint={focusPoint}
+            image={image}
+            imgProps={{ onError }}
+            isFollowingWindowSize={isFollowingWindowSize}
+            size={select(
+                'Size',
+                {
+                    None: undefined,
+                    XXS: Size.xxs,
+                    XS: Size.xs,
+                    S: Size.s,
+                    M: Size.m,
+                    L: Size.l,
+                    XL: Size.xl,
+                    XXL: Size.xxl,
+                },
+                Size.xxl,
+                groupId,
+            )}
+            resizeDebounceTime={resizeDebounceTime}
+            theme={theme}
+            variant={variant}
+        />
+    );
+};
+
+export const brokenThumbnailWithFallbackProps = ({ theme }: StoryProps) => {
+    const align = select<Alignment>('Alignment', Alignment, Alignment.left, groupId);
+    const aspectRatio = select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, groupId);
+    const isCrossOriginEnabled = boolean('Is CORS enabled', false, groupId);
+    const crossOrigin = select('CORS', corsOptions, corsOptions.None, groupId);
+    const fillHeight = boolean('Fill Height', false, groupId);
+    const image = text(
+        'Url',
+        'https://i.picsum.photos/id/237/536/354.jp?hmac=i0yVXW1ORpyCZpQ-CknuyV-jbtU7_x9EBQVhvT5aRr0',
+        groupId,
+    );
+    const variant = select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, groupId);
+
+    return (
+        <Thumbnail
+            align={align}
+            aspectRatio={aspectRatio}
+            isCrossOriginEnabled={isCrossOriginEnabled}
+            crossOrigin={crossOrigin}
+            fillHeight={fillHeight}
+            image={htmlDecode(image)}
+            theme={theme}
+            variant={variant}
+        />
+    );
+};