+++ packages/lumx-react/src/hooks/index.tsx
@@ -17,5 +17,3 @@ export * from './useChipGroupNavigation';
 export * from './useFocus';
 
 export * from './useFocusedImage';
-
-export * from './useImage';