+++ packages/lumx-react/src/components/image-block/ImageBlock.stories.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
-import { Alignment, AspectRatio, Chip, ChipGroup, ImageBlock, Size } from '@lumx/react';
+import { Alignment, AspectRatio, Chip, ChipGroup, CrossOrigin, ImageBlock, Size } from '@lumx/react';
+import { htmlDecode } from '@lumx/react/utils/htmlDecode';
 import { boolean, number, select, text } from '@storybook/addon-knobs';
 
 export default { title: 'LumX components/Image Block' };
@@ -12,12 +13,22 @@ const numberKnobOptions = {
     step: 0.1,
 };
 
+const corsOptions = {
+    None: undefined,
+    Anonymous: CrossOrigin.anonymous,
+    UseCredentials: CrossOrigin.useCredentials,
+};
+
+const groupId = 'Image block';
+
 export const defaultImageBlock = ({ theme }: any) => {
-    const align = select<Alignment>('Alignment', Alignment, Alignment.center, 'Image block');
-    const aspectRatio = select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, 'Image block');
-    const title = text('Title', 'Hello world', 'Image block');
-    const description = text('Description', 'My awesome description', 'Image block');
-    const isDisplayedTags = boolean('Display tags', true, 'Image block');
+    const align = select<Alignment>('Alignment', Alignment, Alignment.center, groupId);
+    const aspectRatio = select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, groupId);
+    const title = text('Title', 'Hello world', groupId);
+    const description = text('Description', 'My awesome description', groupId);
+    const isCrossOriginEnabled = boolean('Is CORS enabled', false, groupId);
+    const crossOrigin = select('CORS', corsOptions, corsOptions.None, groupId);
+    const isDisplayedTags = boolean('Display tags', true, groupId);
     const tags = (
         <ChipGroup align={Alignment.left}>
             <Chip size={Size.s} theme={theme}>
@@ -29,10 +40,14 @@ export const defaultImageBlock = ({ theme }: any) => {
             </Chip>
         </ChipGroup>
     );
-    const imageUrl = text('Url image', 'https://i.picsum.photos/id/1001/2400/1400.jpg', 'Image block');
+    const imageUrl = text(
+        'Url image',
+        'https://i.picsum.photos/id/237/536/354.jpg?hmac=i0yVXW1ORpyCZpQ-CknuyV-jbtU7_x9EBQVhvT5aRr0',
+        groupId,
+    );
     const focusPoint = {
-        x: number('Focus X', 1, numberKnobOptions, 'Image block'),
-        y: number('Focus Y', 0, numberKnobOptions, 'Image block'),
+        x: number('Focus X', 1, numberKnobOptions, groupId),
+        y: number('Focus Y', 0, numberKnobOptions, groupId),
     };
     const size = select(
         'Size',
@@ -42,16 +57,18 @@ export const defaultImageBlock = ({ theme }: any) => {
             XXL: Size.xxl,
         },
         Size.xxl,
-        'Image block',
+        groupId,
     );
 
     return (
         <ImageBlock
             align={align}
             aspectRatio={aspectRatio}
+            crossOrigin={crossOrigin}
             description={description}
             focusPoint={focusPoint}
-            image={imageUrl}
+            image={htmlDecode(imageUrl)}
+            isCrossOriginEnabled={isCrossOriginEnabled}
             size={size}
             tags={isDisplayedTags && tags}
             title={title}