+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Fix `Thumbnail` component loading and error state handling
 
+### Added
+
+-   Add `isCrossOriginEnabled` and `crossOrigin` missing props on `ImageBlock` component to forward them to `Thumbnail` component
+
 ## [0.24.5][] - 2020-08-28
 
 ### Fixed